fix(subagents): satisfy CI after main rebase on PR #136820
- Split settle-wake helpers to stay under max-lines (640 LOC)
- Mock getLatestLiveSubagentRunByChildSessionKey in replay/outbox tests
- Fix retainedMachine locator in cloud-dispatch e2e typecheck
- Add agents-announce tsgo shard; exclude announce tests from agents-other
