+++ test/tsconfig/tsconfig.core.test.agents-other.json
@@ -18,6 +18,8 @@
     "../../src/agents/sandbox/**/*.test.ts",
     "../../src/agents/sandbox/**/*.test.tsx",
     "../../src/agents/tools/**/*.test.ts",
-    "../../src/agents/tools/**/*.test.tsx"
+    "../../src/agents/tools/**/*.test.tsx",
+    "../../src/agents/subagents/announce/**/*.test.ts",
+    "../../src/agents/subagents/announce/**/*.test.tsx"
   ]
 }
